Create A3 Testing and validating
Grant Agreement:

The Activity’s overall objective is to apply and validate the methodology and tools developed under activity 2 to the use case areas that have been selected under the same activity.
Data coming from the various selected sources, including INSPIRE, national data, data from the open source community and other sources will be combined and integrated. 
The new tools developed will then be applied to the data to transform and harmonise it. 
The results and the tools will be validated by the stakeholders involved in the Action in order to check the validity and usability of the resulting information for the selected processes. 

*** Task 3.1	Iterative data processing  (wetransform) *** 
This task will transform the datasets identified under activity 2 into one or more new harmonised data products. For each dataset an iterative approach will be followed, starting from the simplest transformation to immediately getting useful data for validation and feedback (Task 3.3). Then the dataset will be enhanced and made more information-dense according to the needs and the previous results. The resulting products will be subjected to testing, validation, and quality control.
To facilitate effective data transformation and harmonisation novel integration approaches will be used, such as semi-autonomous Extract-Transform-Load (ETL) tools based on machine learning and declarative mapping, unified data management architectures, common in-memory data interfaces, and event-driven data flow architecture.

(D3.1) Report on the iterative processing of all relevant data sets, including formal, interactive documentation for the actual data transformation (ETL) processes, and links to the processed data sets; 

*** Task 3.2	Enhancement of tools (wetransform) *** 
For the ETL processes in task 3.1 the performance will be continuously monitored. To improve performance when processing geographic information, this task will use this information to identify areas of potential performance improvements. Based on the results the developed tools will be enhanced to make the harmonisation process more efficient. It is planned to evaluate cloud processing infrastructures such as ECS (Elastic Cloud Storage), GeoRocket and ElasticSearch for scalability to very high volumes of data, as well as to apply incremental transformation processes for data sets that change often. Furthermore machine learning on cloud resources will be applied specifically to detect when changes in the data would make it necessary to update the ETL process. This can also include making use of cloud GPU resources where this would be effective.
Any findings from the analysis as well as the measured outcomes of the enhancements will be documented and explained in a report. 

(D3.2) Report on issues in performance and other metrics that were identified, and which measures were implemented and evaluated to address these issues:

*** Task 3.3	Documentation, testing and assessment of the tools (geograma) *** 
This task will share the outcomes of the harmonization processes with the relevant stakeholders to get valuable feedback from data providers and users, 
in order to check that the processes are correct and final data is useful for the specific use cases (work processes) identified earlier and 
that it satisfies their data quality requirements. 
In particular it will be made sure that the data and its metadata comply with the European Data Portal standards and that the resulting data sets can be harvested by the European Data Portal. 
The feedback received will be used to further fine-tune the harmonisation tools and transformation processes.

(D3.3) Report on stakeholder feedback and validation results, including validation results from the EDP, as well as key usability KPIs for each version of each data set that is processed